#1650 refactored too big class DefaultWotThingModelValidation in smaller pieces
* strict split between config reading in DefaultWotThingModelValidation and static functionality in Internal* classes
